Move structured text into its own file
