+++ Source/swiftlint/ExplainCommand.swift
@@ -11,42 +11,6 @@ import Commandant
 import LlamaKit
 import SwiftLintFramework
 
-typealias StructuredInlineText = String
-
-extension Array {
-    func dropWhile(x: Element -> Bool) -> [Element] {
-        var count = 0
-        while x(self[count]) { count++ }
-        var copy = self
-        copy.removeRange(0..<count)
-        return copy
-    }
-}
-
-
-extension String {
-    var chomped: String {
-        return String(reverse(self).dropWhile { $0 == "\n" }.reverse())
-    }
-}
-
-enum StructuredText {
-    case Header(level: Int, text: StructuredInlineText)
-    case Paragraph(StructuredInlineText)
-    case List(items: [StructuredText])
-    case Joined([StructuredText])
-    
-    var markdown: String {
-        switch self {
-            
-        case let .Header(level, t): return join("", Array(count: level, repeatedValue: "#")) + " " + t
-        case .Paragraph(let t): return t.chomped
-        case .List(let items): return "\n".join(items.map { "* " + $0.markdown })
-        case .Joined(let items): return "\n\n".join(items.map { $0.markdown } )
-        }
-    }
-}
-
 func describeExample(example: RuleExample) -> StructuredText {
     return .Joined([
         .Header(level: 1, text: example.ruleName),
@@ -68,7 +32,7 @@ struct ExplainCommand: CommandType {
         switch mode {
         case let .Arguments:
             for example: RuleExample in Linter.explainableRules {
-                println(describeExample(example).markdown)
+                println(describeExample(example).ansi)
                                 
             }
